# backend/openapi/dist.yaml
openapi: 3.0.3
info:
  title: Modular JWT Authenticated API
  version: 1.0.0
  description: An example API using JWT imported from a separate file.
servers:
  - url: 'http://localhost:8080/api/v1'
security:
  - BearerAuth: []
paths:
...
  /notes:
    post:
      summary: Create a new note
      operationId: createNote
      tags:
        - Note
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_CreateNote'
      responses:
        '201':
          description: Note created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_DetailNote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /notes/list:
    get:
      summary: List user's notes
      operationId: listNotes
      tags:
        - Note
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: query
          in: query
          schema:
            type: string
          description: Query to search for
        - name: folder_id
          in: query
          schema:
            type: string
          description: Filter notes by folder ID
        - name: status
          in: query
          schema:
            type: string
            enum:
              - draft
              - published
              - archived
          description: Filter notes by status
      responses:
        '200':
          description: Notes retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  notes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Res_DetailNote'
                  total:
                    type: integer
                    description: Total number of notes
                  limit:
                    type: integer
                    description: Number of notes per page
                  offset:
                    type: integer
                    description: Number of notes skipped
                  query:
                    type: string
                    description: Query to search for
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/notes/{note_id}':
    get:
      summary: Get note by ID
      operationId: getNote
      tags:
        - Note
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
      responses:
        '200':
          description: Note retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_DetailNote'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/notes/{note_id}/update':
    put:
      summary: Update note by ID
      operationId: updateNote
      tags:
        - Note
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_UpdateNote'
      responses:
        '200':
          description: Note updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_DetailNote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/notes/{note_id}/delete':
    delete:
      summary: Delete note by ID
      operationId: deleteNote
      tags:
        - Note
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
      responses:
        '200':
          description: Note deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/notes/{note_id}/public-edit':
    get:
      summary: Get public edit settings for a note
      operationId: getPublicEditSettings
      tags:
        - Note
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
      responses:
        '200':
          description: Public edit settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_PublicEditSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Update public edit settings for a note
      operationId: updatePublicEditSettings
      tags:
        - Note
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_PublicEditSettings'
      responses:
        '200':
          description: Updated public edit settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_PublicEditSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/notes/{note_id}/public-edit/rotate':
    post:
      summary: Rotate public edit token for a note
      operationId: rotatePublicEditToken
      tags:
        - Note
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
      responses:
        '200':
          description: Rotated public edit token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_PublicEditSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/notes/{note_id}/tom':
    put:
      summary: Update note top of mind by ID
      operationId: updateNoteTOM
      tags:
        - Note
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
        - name: tom
          in: query
          required: false
          schema:
            type: integer
            nullable: true
            example: 1
          description: Top of mind order (null to unpin)
      responses:
        '200':
          description: Note top of mind updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: '1234567890'
                  top_of_mind:
                    type: integer
                    nullable: true
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /notes/list-tom:
    get:
      summary: List user's top of mind notes
      operationId: listNotesTOM
      tags:
        - Note
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Top of mind notes retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Res_DetailNote'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/public/notes/{note_id}':
    get:
      operationId: getPublicNote
      summary: Get a public note by ID (no auth required)
      description: Fetches a note that has been marked as public. No authentication required.
      security: []
      tags:
        - Note
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: The note ID
      responses:
        '200':
          description: Successful response with note details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_DetailNote'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/public/notes/{note_id}/snapshot':
    post:
      summary: Save HTML snapshot for a note
      operationId: saveNoteSnapshot
      tags:
        - Note
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
        - name: X-Edit-Token
          in: header
          required: false
          schema:
            type: string
          description: Public edit token (optional if Bearer is used)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_NoteSnapshot'
      responses:
        '200':
          description: Snapshot saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_NoteSnapshot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/public/notes/{note_id}/snapshot-tiptap':
    post:
      summary: Save Tiptap snapshot for a note
      operationId: saveNoteTiptapSnapshot
      tags:
        - Note
      security:
        - BearerAuth: []
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
          description: Note ID
        - name: X-Edit-Token
          in: header
          required: false
          schema:
            type: string
          description: Public edit token (optional if Bearer is used)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Req_NoteTiptapSnapshot'
      responses:
        '200':
          description: Tiptap snapshot saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Res_NoteSnapshot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter your JWT token in the format: Bearer <token>.
        The token can be obtained from the `/auth/login` endpoint.
    EditToken:
      type: apiKey
      in: header
      name: X-Edit-Token
      description: Public edit token for note collaboration
  schemas:
...
    Req_CreateNote:
      type: object
      properties:
        title:
          type: string
          example: My Note
        content_type:
          type: string
          example: note/article/image/video/audio
        status:
          type: string
          example: draft/published/archived
        content:
          type: string
          example: This is my note content
        thumbnail:
          type: string
          format: uri
          example: 'https://example.com/thumbnail.png'
        tags:
          type: array
          items:
            type: string
          example:
            - tag1
            - tag2
        folder_id:
          type: string
          nullable: true
          example: folder-uuid-123
        is_public:
          type: boolean
          example: false
      required:
        - title
        - content_type
        - status
        - content
        - thumbnail
        - tags
        - is_public
    Req_UpdateNote:
      type: object
      properties:
        id:
          type: string
          example: '1234567890'
        title:
          type: string
          example: My Note
        content:
          type: string
          example: This is my note content
        content_type:
          type: string
          example: text/markdown/html
        status:
          type: string
          example: draft/published/archived
        thumbnail:
          type: string
          format: uri
          example: 'https://example.com/thumbnail.png'
        tags:
          type: array
          items:
            type: string
          example:
            - tag1
            - tag2
        folder_id:
          type: string
          nullable: true
          example: folder-uuid-123
        is_public:
          type: boolean
          example: false
      required:
        - id
...
    Res_DetailNote:
      type: object
      properties:
        id:
          type: string
          example: '1234567890'
        title:
          type: string
          example: My Note
        content:
          type: string
          example: This is my note content
        content_type:
          type: string
          example: text/markdown/html
        status:
          type: string
          example: draft/published/archived
        top_of_mind:
          type: integer
          nullable: true
          example: 1
        thumbnail:
          type: string
          format: uri
          example: 'https://example.com/thumbnail.png'
        tags:
          type: array
          items:
            type: string
          example:
            - tag1
            - tag2
        folder_id:
          type: string
          nullable: true
          example: folder-uuid-123
        is_public:
          type: boolean
          example: false
        public_edit_enabled:
          type: boolean
          example: false
        public_edit_token:
          type: string
          example: c0e7f6f5f0ff4d1fb2a2c4d09b3bda9e
        created_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2021-01-01T00:00:00Z'
      required:
        - id
        - title
        - content
        - thumbnail
        - tags
        - is_public
        - created_at
        - updated_at
        - content_type
        - status
        - top_of_mind
    Req_PublicEditSettings:
      type: object
      properties:
        enabled:
          type: boolean
          example: true
      required:
        - enabled
    Res_PublicEditSettings:
      type: object
      properties:
        enabled:
          type: boolean
          example: true
        token:
          type: string
          example: c0e7f6f5f0ff4d1fb2a2c4d09b3bda9e
      required:
        - enabled
        - token
    Req_NoteSnapshot:
      type: object
      properties:
        content:
          type: string
          example: <p>Updated note content</p>
      required:
        - content
    Req_NoteTiptapSnapshot:
      type: object
      properties:
        tiptap_content:
          type: string
          example: '{"type":"doc","content":[{"type":"paragraph","content":[{"type":"text","text":"Updated tiptap document"}]}]}'
      required:
        - tiptap_content
    Res_NoteSnapshot:
      type: object
      properties:
        id:
          type: string
          example: note-uuid-123
        updated_at:
          type: string
          format: date-time
          example: '2026-01-24T12:30:00Z'
      required:
        - id
        - updated_at
...
  responses:
    BadRequest:
      description: Bad Request. The request is invalid.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request.
    Unauthorized:
      description: 'Unauthorized. The provided JWT token is missing, invalid, or expired.'
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Authentication token is required.
    NotFound:
      description: Not Found. The requested resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found.
    InternalServerError:
      description: Internal Server Error. The server encountered an unexpected condition that prevented it from fulfilling the request.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Internal server error.
...
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of results to return.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: Number of results to skip for pagination.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0